Move global messages logic to socket listener
