+++ public/js/cat_source/ui.core.js
@@ -307,51 +307,11 @@ window.UI = {
           SegmentActions.updateGlobalWarnings(data.details)
         }
 
-        // check for messages
-        if (data.messages) {
-          var msgArray = $.parseJSON(data.messages)
-          if (msgArray.length > 0) {
-            UI.displayMessage(msgArray)
-          }
-        }
-
         $(document).trigger('getWarning:global:success', {resp: data})
       })
       .catch((errors) => {
         OfflineUtils.failedConnection()
       })
-  },
-  displayMessage: function (messages) {
-    var self = this
-    if ($('body').hasClass('incomingMsg')) return false
-    $.each(messages, function () {
-      var elem = this
-      if (
-        typeof Cookies.get('msg-' + this.token) == 'undefined' &&
-        new Date(this.expire) > new Date() &&
-        typeof self.displayedMessages !== 'undefined' &&
-        self.displayedMessages.indexOf(this.token) < 0
-      ) {
-        var notification = {
-          title: elem.title ? elem.title : 'Notice',
-          text: this.msg,
-          type: elem.level ? elem.level : 'warning',
-          autoDismiss: false,
-          position: 'bl',
-          allowHtml: true,
-          closeCallback: function () {
-            var expireDate = new Date(elem.expire)
-            Cookies.set('msg-' + elem.token, '', {
-              expires: expireDate,
-              secure: true,
-            })
-          },
-        }
-        CatToolActions.addNotification(notification)
-        self.displayedMessages.push(elem.token)
-        return false
-      }
-    })
   },
   registerQACheck: function () {
     clearTimeout(UI.pendingQACheck)
